requirements.txt
certifi==2021.5.30
chardet==4.0.0
filelock==3.0.12
gdown==3.13.0
idna==2.10
PySocks==1.7.1
python-decouple==3.4
requests==2.25.1
six==1.16.0
tqdm==4.61.1
urllib3==1.26.5
zipfile36==0.1.3
speechbrain==0.5.9
mir_eval==0.6
pyloudnorm==0.1.0
neptune-client==0.9.18